# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs MlCity Project",
        "description": "API MlCity",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/sight-types": {
            "get": {
                "tags": [
                    "Sight-type"
                ],
                "summary": "Get all sight type",
                "operationId": "d570f8a7676ffddff030f7f91609a129",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/getTypesId/{id}": {
            "get": {
                "tags": [
                    "Sight-type"
                ],
                "summary": "Get sight type by id",
                "operationId": "79bdc20ce9da863b7700332764ba7cf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/addTypeSight/{sight_id}/{type_id}": {
            "post": {
                "tags": [
                    "Sight-type"
                ],
                "summary": "add type in sight",
                "operationId": "5554ba6a1e6ab0c28f8b0bc4f142f046",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "$sight_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/updateTypeSight/{sight_id}/{type_id}": {
            "put": {
                "tags": [
                    "Sight-type"
                ],
                "summary": "add type in event",
                "operationId": "e974fac5bb43c4dc638295db3fe0311d",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "$sight_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/deleteTypeSight/{sight_id}/{type_id}": {
            "delete": {
                "tags": [
                    "Sight-type"
                ],
                "summary": "add type in sight",
                "operationId": "804a6bb4570139edaa49025ed55d0430",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "$sight_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/statuses": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Get all statuses",
                "operationId": "1976c83876b0986ffe1fe8af73221069",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/getStatusId/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Get statuses by id",
                "operationId": "09fb7418787ce408e160c56a5dc93f36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/addStatusEvent": {
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Add statuses for event",
                "operationId": "91ae16621a82ef1be87f1754e6d14b62",
                "parameters": [
                    {
                        "name": "status_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "event_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/addStatusEvent": {
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Add statuses for sight",
                "operationId": "133afee3dcb80ab83add5690da97031f",
                "parameters": [
                    {
                        "name": "status_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sight_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}